Add files via upload
Jazak-Allah:
This is an Android App for maintaining a database of all the donations been achieved for a mosque. It maintains a database of all the donations, list of the donors,  the expenditures out of donations and keep updating the donors about the expenditures and different programs going on in a particular Mosque. 
App contains an admin panel for members of the Mosque community. Also contains the integration of message service for donors and Community Members.
The Android App includes following:
- Firebase Authentication Integration for Log in/Sign Up Admin Panels
- Firebase Database Integration
- Action Bar Customization
- List View
- SQ Lite Database Integration
- SMS Manager